[fix]rename artgor_utils to lgb_train and remove extra scripts
